+++ src/models/emojis/mod.rs
@@ -1,3 +1,4 @@
+pub mod helpers;
 pub mod queries;
 pub mod types;
 pub mod validators;